+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [7.0.12] - 2023-11-16
+
+- Adds Phone Number normalisation
+
 ## [7.0.11] - 2023-11-10
 
 - Fixes email verification behaviour with user id mapping
