+++ src/test/java/io/supertokens/test/passwordless/api/PasswordlessUserGetAPITest.java
@@ -1,0 +1,94 @@
+/*
+ *    Copyright (c) 2021, VRAI Labs and/or its affiliates. All rights reserved.
+ *
+ *    This software is licensed under the Apache License, Version 2.0 (the
+ *    "License") as published by the Apache Software Foundation.
+ *
+ *    You may not use this file except in compliance with the License. You may
+ *    obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ *    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ *    License for the specific language governing permissions and limitations
+ *    under the License.
+ */
+
+package io.supertokens.test.passwordless.api;
+
+import com.google.gson.JsonObject;
+import io.supertokens.ProcessState;
+import io.supertokens.pluginInterface.STORAGE_TYPE;
+import io.supertokens.pluginInterface.multitenancy.TenantIdentifier;
+import io.supertokens.pluginInterface.passwordless.PasswordlessStorage;
+import io.supertokens.storageLayer.StorageLayer;
+import io.supertokens.test.TestingProcessManager;
+import io.supertokens.test.Utils;
+import io.supertokens.test.httpRequest.HttpRequestForTesting;
+
+import io.supertokens.webserver.WebserverAPI;
+
+import org.junit.AfterClass;
+import org.junit.Before;
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.rules.TestRule;
+
+import java.util.HashMap;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
+
+public class PasswordlessUserGetAPITest {
+  @Rule
+  public TestRule watchman = Utils.getOnFailure();
+
+  @AfterClass
+  public static void afterTesting() {
+    Utils.afterTesting();
+  }
+
+  @Before
+  public void beforeEach() {
+    Utils.reset();
+  }
+
+  @Test
+  public void testGetUserWithUnnormalisedPhoneNumber() throws Exception {
+    String[] args = { "../" };
+
+    TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
+    assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
+
+    if (StorageLayer.getStorage(process.getProcess()).getType() != STORAGE_TYPE.SQL) {
+      return;
+    }
+
+    PasswordlessStorage storage = (PasswordlessStorage) StorageLayer.getStorage(process.getProcess());
+
+    // length of user ID needs to be 36 character long, otherwise it throws error
+    // with postgres DB
+    String userId = "pZ9SP0USbXbejGFO6qx7x3JBjupJZVtw4RkD";
+    String phoneNumber = "+44-207 183 8750";
+    String normalisedPhoneNumber = io.supertokens.utils.Utils.normalizeIfPhoneNumber(phoneNumber);
+
+    storage.createUser(new TenantIdentifier(null, null, null),
+        userId, null, normalisedPhoneNumber, System.currentTimeMillis());
+    {
+      HashMap<String, String> map = new HashMap<>();
+      map.put("phoneNumber", phoneNumber);
+      JsonObject response = HttpRequestForTesting.sendGETRequest(process.getProcess(), "",
+          "http://localhost:3567/recipe/user", map, 1000, 1000, null, WebserverAPI.getLatestCDIVersion().get(),
+          "passwordless");
+
+      assertEquals("OK", response.get("status").getAsString());
+      assert (response.has("user"));
+      JsonObject user = response.get("user").getAsJsonObject();
+      assertEquals(user.get("phoneNumbers").getAsJsonArray().get(0).getAsString(), normalisedPhoneNumber);
+    }
+
+    process.kill();
+    assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
+  }
+
+}
